requirements.txt
pydub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==20.8
openai-whisper==20231117

